+++ src/app/events/events.module.ts
@@ -4,11 +4,13 @@ import { CommonModule } from '@angular/common';
 import { EventsRoutingModule } from './events-routing.module';
 import { EventsComponent } from './containers/events/events.component';
 import { MdToHtmlPipe } from '../pipes/md-to-html.pipe';
+import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
   imports: [
     CommonModule,
-    EventsRoutingModule
+    EventsRoutingModule,
+    SharedModule
   ],
   exports: [
     EventsComponent