+++ src/app/shared/shared.module.ts
@@ -3,6 +3,8 @@ import { CommonModule } from '@angular/common';
 import { TranslateModule } from '@ngx-translate/core';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { FormsModule } from '@angular/forms';
+import { PageHeaderComponent } from './components/page-header/page-header.component';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 @NgModule({
   imports: [
@@ -10,13 +12,16 @@ import { FormsModule } from '@angular/forms';
     TranslateModule,
     FlexLayoutModule,
     FormsModule,
+    FontAwesomeModule
   ],
   exports: [
     CommonModule,
     TranslateModule,
     FlexLayoutModule,
     FormsModule,
+    PageHeaderComponent,
+    FontAwesomeModule
   ],
-  declarations: []
+  declarations: [ PageHeaderComponent ]
 })
 export class SharedModule { }