+++ src/app/team/team.module.ts
@@ -3,12 +3,17 @@ import { CommonModule } from '@angular/common';
 
 import { TeamRoutingModule } from './team-routing.module';
 import { TeamComponent } from './containers/team/team.component';
+import { SharedModule } from '../shared/shared.module';
+import { TeamMemberComponent } from './components/team-member/team-member.component';
+import { MaterialModule } from '../material/material.module';
 
 @NgModule({
   imports: [
     CommonModule,
-    TeamRoutingModule
+    TeamRoutingModule,
+    SharedModule,
+    MaterialModule
   ],
-  declarations: [TeamComponent]
+  declarations: [TeamComponent, TeamMemberComponent]
 })
 export class TeamModule { }